Rename date_time_formats as date_time_defaults
